+++ components/PageLoading.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import Loading from '../styles/animations/json/loading'
+
+function PageLoading () {
+  return (
+    <Loading />
+  )
+}
+
+export default PageLoading
