+++ pages/index.js
@@ -3,6 +3,7 @@ import Layout from '../components/Layout'
 import Search from '../components/Search'
 import Hero from '../components/Hero'
 import Card from '../components/Card'
+import PageLoading from '../components/PageLoading'
 
 import { useState, useEffect } from 'react'
